Renaming shared package eslint-config-custom to eslint-config
