+++ apps/api/package.json
@@ -40,7 +40,7 @@
   "devDependencies": {
     "@jest/globals": "^29.7.0",
     "eslint": "^9.32.0",
-    "eslint-config-custom": "*",
+    "eslint-config": "*",
     "@types/bun": "^1.1.6",
     "@types/jsonld": "^1.5.14",
     "@types/lodash": "^4.17.6"
