+++ apps/docs/package.json
@@ -36,7 +36,7 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "eslint": "^9.32.0",
-    "eslint-config-custom": "*",
+    "eslint-config": "*",
     "postcss": "^8.5.6",
     "tailwindcss": "^4.1.11",
     "tsconfig": "*",
