+++ apps/docs/src/content/nb/developers/monorepo.mdx
@@ -123,28 +123,28 @@ Denne grafen ble generert av Turborepo 25.04.23.
 ```mermaid
 graph TD
 	CXTX("___ROOT___#build") --> MKUC("___ROOT___")
-	XPYH("api#build") --> TLTN("eslint-config-custom#build")
+	XPYH("api#build") --> TLTN("eslint-config#build")
 	XPYH("api#build") --> LGSI("utils#build")
-	CGWW("chc-exh-nt-web#build") --> TLTN("eslint-config-custom#build")
+	CGWW("chc-exh-nt-web#build") --> TLTN("eslint-config#build")
 	CGWW("chc-exh-nt-web#build") --> AHYC("tailwind-config#build")
 	CGWW("chc-exh-nt-web#build") --> FPMW("tailwind-ui#build")
 	CGWW("chc-exh-nt-web#build") --> YQKY("tsconfig#build")
-	STSI("chc-web#build") --> TLTN("eslint-config-custom#build")
+	STSI("chc-web#build") --> TLTN("eslint-config#build")
 	STSI("chc-web#build") --> YQKY("tsconfig#build")
 	MDNO("design-system#build") --> CXTX("___ROOT___#build")
-	FNHG("docs#build") --> TLTN("eslint-config-custom#build")
+	FNHG("docs#build") --> TLTN("eslint-config#build")
 	FNHG("docs#build") --> FPMW("tailwind-ui#build")
 	FNHG("docs#build") --> YQKY("tsconfig#build")
-	TLTN("eslint-config-custom#build") --> CXTX("___ROOT___#build")
-	BFCM("link-shortener#build") --> TLTN("eslint-config-custom#build")
+	TLTN("eslint-config#build") --> CXTX("___ROOT___#build")
+	BFCM("link-shortener#build") --> TLTN("eslint-config#build")
 	BFCM("link-shortener#build") --> YQKY("tsconfig#build")
-	IGME("sps-termportal-web#build") --> TLTN("eslint-config-custom#build")
+	IGME("sps-termportal-web#build") --> TLTN("eslint-config#build")
 	AHYC("tailwind-config#build") --> CXTX("___ROOT___#build")
-	FPMW("tailwind-ui#build") --> TLTN("eslint-config-custom#build")
+	FPMW("tailwind-ui#build") --> TLTN("eslint-config#build")
 	FPMW("tailwind-ui#build") --> AHYC("tailwind-config#build")
 	FPMW("tailwind-ui#build") --> YQKY("tsconfig#build")
 	YQKY("tsconfig#build") --> CXTX("___ROOT___#build")
-	LGSI("utils#build") --> TLTN("eslint-config-custom#build")
+	LGSI("utils#build") --> TLTN("eslint-config#build")
 	LGSI("utils#build") --> YQKY("tsconfig#build")
 ```
