+++ apps/sps-stadnamn-web/package.json
@@ -29,7 +29,7 @@
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
-    "eslint-config-custom": "*",
+    "eslint-config": "*",
     "eslint-config-next": "14.1.0",
     "next-plausible": "^3.12.0",
     "postcss": "^8",
