+++ apps/sps-termportal-web/package.json
@@ -30,7 +30,7 @@
     "@vueuse/core": "^10.5.0",
     "@vueuse/nuxt": "^10.5.0",
     "eslint": "^9.0.0",
-    "eslint-config-custom": "*",
+    "eslint-config": "*",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-tsdoc": "^0.2.17",
     "jsonld": "^8.1.0",
