+++ packages/eslint-config/nuxt.js
