+++ packages/eslint-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "eslint-config-custom",
+  "name": "eslint-config",
   "version": "1.0.0",
   "license": "MIT",
   "type": "module",
