+++ packages/eslint-config/react.js
