+++ packages/utils/package.json
@@ -14,7 +14,7 @@
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
   },
   "devDependencies": {
-    "eslint-config-custom": "*",
+    "eslint-config": "*",
     "tsconfig": "*",
     "tsup": "^8",
     "typescript": "^4.9.3"
